+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [?NEXT?] - ????-??-??
 
-- TODO: write changelog before preparing next tag
+- Improve rerendering of the `BootstrapProvider`.
 
 ## [v1.5.14] - 2018-04-12